Fixes #366: Core and Module improvements
+Core:
Introduces a migration object to manage all database conversion options and handle the conversion in a generic way.
Introduces FilterModules which may be inserted between the import and export, and through which all data is passed.
Introduces an ObservableModule, which is a filter module that supports adding observers for different database migration steps.
Refactors main logging operations to be done by the 'ProgressLoggerObserver', which can be attached to an ObservableModule to provide logging functionality (or completely disabled if logging is not needed).
Code format and cleanup (like changing string literals into constants).
Improvements to CLI arguments and error handling and module help text display.
Changed configuration folder structure: reports and logs are saved in the same folder as the dbptk jar file, unless the property dbptk.home is used to specify a different directory.

+Modules (Oracle):
Renamed 'database' to 'instance' in the oracle modules, as that seems to be the correct name in Oracle.

+Modules (Dbvtk):
Rename 'Solr' module to 'Dbvtk' module (later to be moved from DBPTK project into DBVTK project).
Dbvtk module requires the LOB output folder to be passed in as a parameter.
[WIP] Refactored Dbvtk module to use 1 solr collection per database (previously 1 collection was created for each table).

+Modules (SIARD):
[WIP] Began extending SIARD 2 into SIARD 2.0 and SIARD 2.1.